rework to use service resource directly instead of subclassing
