+++ lib/chef/provider/service/systemd.rb
@@ -1,5 +1,6 @@
 #
 # Author:: Stephen Haynes (<sh@nomitor.com>)
+# Author:: Davide Cavalca (<dcavalca@fb.com>)
 # Copyright:: Copyright 2011-2016, Chef Software Inc.
 # License:: Apache License, Version 2.0
 #
@@ -32,13 +33,6 @@ class Chef::Provider::Service::Systemd < Chef::Provider::Service::Simple
 
   def self.supports?(resource, action)
     Chef::Platform::ServiceHelpers.config_for_service(resource.service_name).include?(:systemd)
-  end
-
-  def load_new_resource_state
-    super
-    if ( @new_resource.masked.nil? )
-      @new_resource.masked(@current_resource.masked)
-    end
   end
 
   def load_current_resource
@@ -74,32 +68,6 @@ class Chef::Provider::Service::Systemd < Chef::Provider::Service::Simple
       a.whyrun ["Failed to determine status of #{new_resource}, using command #{new_resource.status_command}.",
         "Assuming service would have been installed and is disabled"]
     end
-  end
-
-  def action_mask
-    if @current_resource.masked
-      Chef::Log.debug("#{@new_resource} already masked - nothing to do")
-    else
-      converge_by("mask service #{@new_resource}") do
-        mask_service
-        Chef::Log.info("#{@new_resource} masked")
-      end
-    end
-    load_new_resource_state
-    @new_resource.masked(true)
-  end
-
-  def action_unmask
-    if @current_resource.masked
-      converge_by("unmask service #{@new_resource}") do
-        unmask_service
-        Chef::Log.info("#{@new_resource} masked")
-      end
-    else
-      Chef::Log.debug("#{@new_resource} already unmasked - nothing to do")
-    end
-    load_new_resource_state
-    @new_resource.masked(false)
   end
 
   def start_service
